+++ src/CleanRegex/Internal/Pcre/DeprecatedMatchDetail.php
@@ -17,11 +17,11 @@ class DeprecatedMatchDetail
     /**
      * @deprecated
      */
-    public static function create(Subject              $subject, int $index, int $limit,
+    public static function create(Subject              $subject, int $index,
                                   IRawMatchOffset      $match, MatchAllFactory $allFactory,
                                   GroupFactoryStrategy $strategy = null): MatchDetail
     {
-        return new MatchDetail($subject, $index, $limit, $match, $match, $match, $match, $allFactory,
+        return new MatchDetail($subject, $index, $match, $match, $match, $match, $allFactory,
             $strategy ?? new MatchGroupFactoryStrategy(),
             new PerformanceSignatures($match, $match));
     }