+++ src/CleanRegex/Match/Details/Detail.php
@@ -22,12 +22,6 @@ interface Detail extends Structure, Intable
     public function index(): int;
 
     /**
-     * @return int
-     * @deprecated
-     */
-    public function limit(): int;
-
-    /**
      * @param string|int $nameOrIndex
      * @return string
      */